+++ app/controllers/organen/orgaan/index.js
@@ -17,4 +17,16 @@ export default class OrganenOrgaanIndexController extends Controller {
     this.isModalActive = !this.isModalActive;
     setTimeout(() => this.router.refresh(), 1000);
   }
+
+  @action
+  async archiveOrgaan() {
+    this.model.bestuursorgaan.deactivatedAt = new Date();
+    await this.model.bestuursorgaan.save();
+  }
+
+  @action
+  async deArchiveOrgaan() {
+    this.model.bestuursorgaan.deactivatedAt = undefined;
+    await this.model.bestuursorgaan.save();
+  }
 }
