+++ app/templates/organen/orgaan.hbs
@@ -21,50 +21,7 @@
   </main.sidebar>
   <main.content>
     <AuBodyContainer id="content">
-      <AuToolbar @border="bottom" @size="large" as |Group|>
-        <Group>
-          <AuHeading @skin="2">
-            {{this.model.bestuursorgaan.naam}}</AuHeading>
-          <AuPill
-            @skin="link"
-            @draft={{false}}
-            @size="regular"
-            @icon="info-circle"
-            @iconAlignment="left"
-            @hideText={{false}}
-          >
-            {{if
-              (await this.model.bestuursorgaan.isDecretaal)
-              "Decretaal"
-              "Niet decretaal"
-            }}
-          </AuPill>
-        </Group>
-        <Group>
-          <AuButtonGroup>
-            {{#unless (await @model.bestuursorgaan.isDecretaal)}}
-              {{#if @model.bestuursorgaan.isActive}}
-                <AuButton
-                  @icon="archive"
-                  @skin="secondary"
-                  @alert="true"
-                  {{on "click" this.archiveOrgaan}}
-                >Archiveer</AuButton>
-              {{else}}
-                <AuButton
-                  @icon="undo"
-                  @skin="secondary"
-                  @alert="true"
-                  {{on "click" this.deArchiveOrgaan}}
-                >Heractiveer</AuButton>
-              {{/if}}
-            {{/unless}}
-          </AuButtonGroup>
-        </Group>
-      </AuToolbar>
-      <div class="au-c-body-container au-c-body-container--scroll">
-        {{outlet}}
-      </div>
+      {{outlet}}
     </AuBodyContainer>
   </main.content>
 </AuMainContainer>
